+++ package.json
@@ -7,7 +7,7 @@
     "build": "webpack --mode production",
     "lint": "npm-run-all lint:*",
     "lint:ts": "eslint --ext=.ts,.tsx src --fix",
-    "lint:css": "stylelint src/**/*.css --fix",
+    "lint:css": "stylelint src/**//**/*.css --fix",
     "prepare": "husky install"
   },
   "repository": {
